+++ EQLogParser/src/ui/chart/HealingColumnChart.cs
@@ -1,0 +1,9 @@
+﻿namespace EQLogParser
+{
+  internal class HealingColumnChart : ColumnChart
+  {
+    protected override IStatsManager StatsManager => HealingStatsManagerAdapter.Instance;
+
+    protected override string ChartTitle => "Players vs Top Performer (Percent of Total Healing)";
+  }
+}